+++ pkg/proxy/base.go
@@ -95,6 +95,8 @@ func ParseProxyFromLink(link string) (p Proxy, err error) {
 		p, err = ParseTrojanLink(link)
 	} else if strings.HasPrefix(link, "vless://") {
 		p, err = ParseVlessLink(link)
+	} else if strings.HasPrefix(link, "anytls://") {
+		p, err = ParseAnyTLSLink(link)
 	}
 	if err != nil || p == nil {
 		return nil, errors.New("link parse failed")
@@ -145,6 +147,13 @@ func ParseProxyFromClashProxy(p map[string]interface{}) (proxy Proxy, err error)
 		return &proxy, nil
 	case "vless":
 		var proxy Vless
+		err := json.Unmarshal(pjson, &proxy)
+		if err != nil {
+			return nil, err
+		}
+		return &proxy, nil
+	case "anytls":
+		var proxy AnyTLS
 		err := json.Unmarshal(pjson, &proxy)
 		if err != nil {
 			return nil, err